Update 1-print_binary.c -f

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -14,7 +14,7 @@ void print_binary(unsigned long int n)
 		_putchar('0');
 		return;
 	}
-	while (i < (sizeof(unsigned int) * 8))
+	while (i < (sizeof(unsigned long int) * 8))
 	{
 		if ((default_n << i) > n)
 		{
